feat(rpc): add eth_pubsub skeleton
